ref: Add traits for Checker and ResultsProducer
